ackermann: add implementations of ackermann function
The ackermann function has been implemented in two ways,

- straight forward recursion
- memoized version (slightly optimizes number of calls)

It's not surprising that the ackermann function's memoized version
didn't help much in computing values for larger parameters. This
is because of its nature of having fewer recurring subproblems.
This shows how memoization isn't of much use when there are fewer
recurring subproblems.