test: magic method invoked when subclasses implement them differently
This test had a base class with __lt__ method implemented in one way.
The 2 subclasses implemented them each in a different way. When an
object of each subclass was used to compare the method invoked
depended on which object was to the left of the '<' operator (or to the
right of the '>' operator). This behaviour is partly expected but it's
surprising (but sounds natural) that even though __gt__ has not been
implemented s1 > s2 behaves the same as s2 < s1.

This shows that,

  - method of base class will be invoked only when object to left of
    '<' operator doesn't implement override __lt__ and when called
    explicitly

  - when overriding a base class method the subclass should ensure the
    operands are of the expected types

  - if __gt__ is not defined but __lt__ is defined then the expression
    s1 > s2 results in __lt__ being called (as if the expression was
    s2 < s1) and vice versa.